add chinese_simpleqa config
